Update of the TobiiProGlasses2 controler for video and gaze streaming
