+++ examples/streaming.py
@@ -15,7 +15,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>
 
-from tobiiglassesctrlupdate import TobiiGlassesController
+from tobiiglassesctrl import TobiiGlassesController
 
 import sys
 import time
